requirements.txt
torch==1.1.0
torchvision==0.3.0
Pillow==6.1.0
tensorboardX==1.8
ptflops==0.4
matplotlib
lmdb
tensorboard



